Move initWorkerModule to its own file like setupEventStream
