+++ src/Worker.ts
@@ -5,6 +5,7 @@ import * as parseArgs from 'minimist';
 import { FunctionLoader } from './FunctionLoader';
 import { CreateGrpcEventStream } from './GrpcClient';
 import { setupEventStream } from './setupEventStream';
+import { setupWorkerModule } from './setupWorkerModule';
 import { ensureErrorType } from './utils/ensureErrorType';
 import { InternalException } from './utils/InternalException';
 import { systemError, systemLog } from './utils/Logger';
@@ -42,6 +43,7 @@ export function startNodeWorker(args) {
 
     const channel = new WorkerChannel(eventStream, new FunctionLoader());
     setupEventStream(workerId, channel);
+    setupWorkerModule(channel);
 
     eventStream.write({
         requestId: requestId,
