+++ test/eventHandlers/beforeEventHandlerSuite.ts
@@ -4,6 +4,7 @@
 import * as sinon from 'sinon';
 import { FunctionLoader } from '../../src/FunctionLoader';
 import { setupEventStream } from '../../src/setupEventStream';
+import { setupWorkerModule } from '../../src/setupWorkerModule';
 import { WorkerChannel } from '../../src/WorkerChannel';
 import { TestEventStream } from './TestEventStream';
 
@@ -12,5 +13,6 @@ export function beforeEventHandlerSuite() {
     const loader = sinon.createStubInstance<FunctionLoader>(FunctionLoader);
     const channel = new WorkerChannel(stream, loader);
     setupEventStream('workerId', channel);
+    setupWorkerModule(channel);
     return { stream, loader, channel };
 }
